+++ examples/web/src/pages/StartPage.vue
@@ -47,6 +47,12 @@ import PageLink from '../components/PageLink.vue'
           API documentation within a pre-existing website.
         </template>
       </PageLink>
+      <PageLink to="path-routing-api-reference">
+        <template #title>API Reference with Path Routing</template>
+        <template #description>
+          Standalone API Reference with path routing instead of hash routing
+        </template>
+      </PageLink>
     </div>
     <h1>@scalar/api-client</h1>
     <div class="page-links">
@@ -60,6 +66,10 @@ import PageLink from '../components/PageLink.vue'
     </div>
     <h1>Examples</h1>
     <div class="page-links">
+      <PageLink href="http://localhost:5063/scalar">
+        <template #title>Docusaurus</template>
+        <template #description>@scalar/docusaurus</template>
+      </PageLink>
       <PageLink href="http://localhost:5058">
         <template #title>Next.js</template>
         <template #description>@scalar/nextjs-api-reference</template>