+++ packages/api-reference/src/components/Content/Introduction/Description.vue
@@ -2,13 +2,13 @@
 import {
   type DescriptionSectionSSRKey,
   type SSRState,
+  createHash,
   ssrState,
 } from '@scalar/oas-utils'
 import { computedAsync } from '@vueuse/core'
 import { onServerPrefetch, useSSRContext } from 'vue'
 
 import {
-  createHash,
   getHeadingsFromMarkdown,
   getLowestHeadingLevel,
   sleep,
@@ -51,22 +51,31 @@ const sections = computedAsync(
   ssrState[ssrStateKey] ?? [], // initial state
 )
 
-const { getHeadingId, hash, isIntersectionEnabled } = useNavState()
+const { getHeadingId, hash, isIntersectionEnabled, pathRouting } = useNavState()
 
-function handleScroll(headingId: string) {
+function handleScroll(headingId = '') {
   if (!isIntersectionEnabled.value) return
+
+  const newUrl = new URL(window.location.href)
+
+  // If we are pathrouting, set path instead of hash
+  if (pathRouting.value) {
+    newUrl.pathname = pathRouting.value.basePath + '/' + headingId
+  } else {
+    newUrl.hash = headingId
+  }
+  hash.value = headingId
 
   // We use replaceState so we don't trigger the url hash watcher and trigger a scroll
   // this is why we set the hash value directly
-  window.history.replaceState({}, '', `#${headingId}`)
-  hash.value = headingId ?? ''
+  window.history.replaceState({}, '', newUrl)
 }
 
 // SSR hack - waits for the computedAsync to complete then we save the state
 onServerPrefetch(async () => {
   const ctx = useSSRContext<SSRState>()
   await sleep(1)
-  ctx!.scalarState[ssrStateKey] = sections.value
+  ctx!.payload.data[ssrStateKey] = sections.value
 })
 </script>
 <template>