+++ packages/api-reference/src/hooks/useNavState.ts
@@ -1,12 +1,18 @@
-import type { Heading, TransformedOperation } from '@scalar/oas-utils'
+import {
+  type Heading,
+  type TransformedOperation,
+  ssrState,
+} from '@scalar/oas-utils'
 import { slug } from 'github-slugger'
-import { onMounted, ref } from 'vue'
+import { ref } from 'vue'
 
-import { scrollToId, sleep } from '../helpers'
-import type { Tag } from '../types'
+import type { PathRouting, Tag } from '../types'
 
 // Keeps track of the URL hash without the #
-const hash = ref('')
+const hash = ref(ssrState.hash ?? '')
+
+// Are we using path routing
+const pathRouting = ref<PathRouting | undefined>()
 
 // To disable the intersection observer on click
 const isIntersectionEnabled = ref(false)
@@ -20,6 +26,13 @@ const getHeadingId = (heading: Heading) => {
   }
 
   return ''
+}
+
+const getPathRoutingId = (pathName: string) => {
+  if (!pathRouting.value) return ''
+
+  const reggy = new RegExp('^' + pathRouting.value?.basePath + '/?')
+  return decodeURIComponent(pathName.replace(reggy, ''))
 }
 
 const getWebhookId = (name?: string, httpVerb?: string) => {
@@ -61,24 +74,11 @@ const getSectionId = (hashStr = hash.value) => {
 }
 
 // Update the reactive hash state
-const updateHash = () => (hash.value = window.location.hash.replace(/^#/, ''))
-
-// We should call this as little as possible, ideally once
-const enableHashListener = () =>
-  onMounted(async () => {
-    updateHash()
-    window.onhashchange = async () => {
-      isIntersectionEnabled.value = false
-      updateHash()
-
-      // TODO: we should be able to remove this once we find the cause
-      // for some reason pressing back doesn't always scroll to the correct section
-      scrollToId(window.location.hash.replace(/^#/, ''))
-
-      await sleep(100)
-      isIntersectionEnabled.value = true
-    }
-  })
+const updateHash = () => {
+  hash.value = pathRouting.value
+    ? getPathRoutingId(window.location.pathname)
+    : decodeURIComponent(window.location.hash.replace(/^#/, ''))
+}
 
 /**
  * Hook which provides reactive hash state from the URL
@@ -86,20 +86,17 @@ const enableHashListener = () =>
  *
  * isIntersectionEnabled is a hack to prevent intersection observer from triggering
  * when clicking on sidebar links or going backwards
- *
- *
- * @param hasLifecyle - we cannot use lifecycle hooks when called from another composable, this prevents that
  */
-export const useNavState = () => {
-  return {
-    hash,
-    getWebhookId,
-    getModelId,
-    getHeadingId,
-    getOperationId,
-    getSectionId,
-    getTagId,
-    isIntersectionEnabled,
-    enableHashListener,
-  }
-}
+export const useNavState = () => ({
+  hash,
+  getWebhookId,
+  getModelId,
+  getHeadingId,
+  getOperationId,
+  getPathRoutingId,
+  getSectionId,
+  getTagId,
+  isIntersectionEnabled,
+  pathRouting,
+  updateHash,
+})