+++ packages/docusaurus/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @scalar/docusaurus
 
+## 0.1.36
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.36
+
+## 0.1.35
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.35
+
+## 0.1.34
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.34
+
+## 0.1.33
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.33
+
+## 0.1.32
+
+### Patch Changes
+
+- @scalar/api-reference-react@0.1.32
+
 ## 0.1.31
 
 ### Patch Changes